Move logging test classes to where they belong.
