+++ src/test/java/org/betonquest/betonquest/modules/logger/handler/chat/PlayerPackageReceiverSelectorTest.java
@@ -1,7 +1,6 @@
-package org.betonquest.betonquest.modules.logger.handler;
+package org.betonquest.betonquest.modules.logger.handler.chat;
 
 import org.betonquest.betonquest.modules.logger.BetonQuestLogRecord;
-import org.betonquest.betonquest.modules.logger.handler.chat.PlayerPackageReceiverSelector;
 import org.betonquest.betonquest.modules.logger.util.BetonQuestLoggerService;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
