+++ src/test/java/org/betonquest/betonquest/modules/logger/handler/history/QueueBackedLogRecordQueueTest.java
@@ -1,6 +1,5 @@
-package org.betonquest.betonquest.modules.logger.handler;
+package org.betonquest.betonquest.modules.logger.handler.history;
 
-import org.betonquest.betonquest.modules.logger.handler.history.QueueBackedLogRecordQueue;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
